Remove abstract method
